requirements.txt
streamlit
opencv-python
numpy
websocket-client
ultralytics
fastapi
uvicorn[standard]
rclpy
rosidl-runtime-py
pyserial